classes and methods
learning classes in python. Classes are a mix of attributes and methods. Methods are functions that belong to a class. The 'self' keyword represents the instance of the class, it allows us to access the attributes and methods of that class. It can be called by any name (like 'instance') but self is commonly used and thus makes our code easier to read. We also created methods with exta parameters that would return based on booleans.

Finally and perhaps most importantly we used the __init__ method. It is a constructor in OO terminology. This method is called when an object is created from the class and initalises the said attributes from the class (inc. kwargs).

We also used a for loop (self, key, value) and 'setattr' for what I believe is a way to pair keys with values.