Flask - Static Files
Beyond a template, we might want to do some styling, or add javascript. So we created a '/static' directory for flask. Then we created a style.css file and a scripts.js file. After we added our standard css and js code, we linked these static files into our /templates through layout.html. Like usual, we added the <link> for styling in the title block and the <script> just before the body closing.