Diary App (WIP)
In this short program, we begin by importing a python collection (ordered dict) along with ORM (peewee) and datetime. As usual we then create a database, create a clause to enact that database and create an Entry class that inherits from Model (the class is also named as a singular item 'Entry' as it represents a single item. We then define the content as a TextField() and create a timestamp variable. Inside this class we also define a Meta class as a means of telling the model which database it belongs to.

Menu is an ordered dictionary that pairs values (soon to be input by users) with functions (defined in the completed app).

Initalize function connects to our database and creates a table.

Menu_loop function loops through our ordered dictionary to present the options to the user.

We also define other functions for later:

add_entry
view_entries
delete_entry